ExRequestItemAutoPeel total peel count is a long number.

+++ L2J_Mobius_10.3_MasterClass/java/org/l2jmobius/gameserver/network/clientpackets/autopeel/ExRequestItemAutoPeel.java
@@ -42,7 +42,7 @@ public class ExRequestItemAutoPeel implements ClientPacket
 	public void read(ReadablePacket packet)
 	{
 		_itemObjectId = packet.readInt();
-		_totalPeelCount = packet.readInt();
+		_totalPeelCount = packet.readLong();
 		_remainingPeelCount = packet.readLong();
 	}
